integrador-back/requirements.txt
asgiref==3.8.1
class-registry==2.1.2
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
django-role-permissions==3.2.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
filters==1.3.2
numpy==2.2.4
openpyxl==3.1.5
pandas==2.2.3
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2024.11.6
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2